% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HMRC_API.R
\name{add_region_name}
\alias{add_region_name}
\title{Region tidier}
\usage{
add_region_name(df)
}
\arguments{
\item{df}{any dataframe that contains a GovRegionId column}
}
\description{
Used by all RTS calls to add on the correct region name.
}
\examples{
 add_region_name(df)
}
\keyword{API}
\keyword{HMRC}
