% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HMRC_API.R
\name{call_sitc_hmrc_ots_with_product_code}
\alias{call_sitc_hmrc_ots_with_product_code}
\title{HMRC OTS query base on code and SITC}
\usage{
call_sitc_hmrc_ots_with_product_code(
  time_window = NULL,
  filter_for_code = NULL,
  country_filter = NULL,
  flow_filter = NULL,
  groupby = NULL
)
}
\arguments{
\item{time_window}{like "MonthId gt 202000 and MonthId lt 202100" which is everything from 2020 or "MonthId gt 201908 and MonthId lt 202009" last 12 month up to end of August}

\item{filter_for_code}{A SITC commodity code e.g. "00" or "001300"}

\item{country_filter}{use ISO-2-alpha code e.g. "FR" check HMRC for inconsistencies such as former Yugoslavian countries.}

\item{flow_filter}{choose from "export" "import" or "EU"}

\item{groupby}{takes a vector of keywords the full range is  c("country", "year", "month", "SITC1", "SITC2", "SITC3", "SITC4")}
}
\description{
Querying the HMRC OTS  data using SITC nomenclature. The arguments are not case-sensitive the API itself is still in BETA thus we continue reporting issues.
}
\examples{
 call_sitc_hmrc_ots_with_product_code(filter_for_code = "02", time_window = "MonthId eq 201901", groupby = c("sitc4", "COUNTRY"))
 call_sitc_hmrc_ots_with_product_code(time_window = "MonthId gt 201906 and MonthId lt 202008", country_filter = "FR", flow_filter = "EXport", groupby = c("SITc2", "COUNTRY"))
}
