% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HMRC_API.R
\name{get_eu_estimates}
\alias{get_eu_estimates}
\title{HMRC OTS EU estimates corrector}
\usage{
get_eu_estimates(
  time_window,
  code_filter = NULL,
  flow_filter = "eu",
  code_type,
  groupby
)
}
\arguments{
\item{time_window}{like "MonthId gt 202000 and MonthId lt 202100" which is everything from 2020 or "MonthId gt 201908 and MonthId lt 202009" last 12 month up to end of August. It is taken automatically from the query}

\item{code_filter}{like "02" only works with HS as there is no breakdown for SITC on the API. It is taken automatically from the query}

\item{flow_filter}{choose from "export" "import" or "EU"}

\item{code_type}{'hs' or 'sitc' populated automatically in code}

\item{groupby}{takes a vector of keywords the full range is  c("country", "year", "month", "hs2"). Disaggregation for EU estimates is not available at a lower level or at SITC level. Taken automatically from function.}
}
\description{
Mostly used to add EU estimates corrections to queries but can be used on its own
}
\examples{
 get_eu_estimates(time_window = "MonthId gt 201906 and MonthId lt 202008", code_type = 'hs', groupby = c("country", "year", "HS2"))
 get_eu_estimates(time_window = "MonthId gt 201906 and MonthId lt 202008", code_filter='02', code_type = 'hs', groupby = c("month", "HS2"))
 get_eu_estimates(time_window = "MonthId gt 202008", code_type = 'sitc', groupby = c("country", "year", "month"))
}
\keyword{API}
\keyword{HMRC}
