% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HMRC_API.R
\name{hmrc_api_requester}
\alias{hmrc_api_requester}
\title{HMRC api query function}
\usage{
hmrc_api_requester(full_url)
}
\description{
Used by all wrapper functions it takes the URL replaces the spaces with '\%20' so the query doesn't break and then tidies the output getting rid of unnecessary things. Gives you a warning if your query is too big.
}
\examples{
 hmrc_api_requester("https://api.uktradeinfo.com/ots?$apply=filter(MonthId gt 202008 and CommoditySitcId lt -1)&$count=true")
}
\keyword{API}
\keyword{HMRC}
