% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmrcwrappeR-package.R
\docType{package}
\name{hmrcwrappeR-package}
\alias{hmrcwrappeR}
\alias{hmrcwrappeR-package}
\title{hmrcwrappeR: Mostly painless API calls to the new HMRC API}
\description{
This package was developed for the typical one-off HMRC requests done in DIT as a quick access tool. The tool is under development resolving issues with the API as they show up. If you have any requests please email me. To find out what functions are in the pacakage run \code{ls("package:hmrcwrappeR")}.
}
\author{
\strong{Maintainer}: Fruzsina Gajdos \email{Fruzsina.Gajdos@trade.gov.uk}

Authors:
\itemize{
  \item Kieran Driscoll \email{Kieran.Driscoll@trade.gov.uk}
}

}
\keyword{internal}
